refactor: split FileViewer into content-type specific viewers and reorganize in file-viewer directory
